+++ lib/fuse.c
@@ -198,7 +198,9 @@ struct fuse_context_i {
 
 /* Defined by FUSE_REGISTER_MODULE() in lib/modules/subdir.c and iconv.c.  */
 extern fuse_module_factory_t fuse_module_subdir_factory;
+#ifdef HAVE_ICONV
 extern fuse_module_factory_t fuse_module_iconv_factory;
+#endif
 
 static pthread_key_t fuse_context_key;
 static pthread_mutex_t fuse_context_lock = PTHREAD_MUTEX_INITIALIZER;
@@ -889,6 +891,36 @@ static struct node *find_node(struct fuse *f, fuse_ino_t parent,
 out_err:
 	pthread_mutex_unlock(&f->lock);
 	return node;
+}
+
+static int lookup_path_in_cache(struct fuse *f,
+		const char *path, fuse_ino_t *inop)
+{
+	char *tmp = strdup(path);
+	if (!tmp)
+		return -ENOMEM;
+
+	pthread_mutex_lock(&f->lock);
+	fuse_ino_t ino = FUSE_ROOT_ID;
+
+	int err = 0;
+	char *save_ptr;
+	char *path_element = strtok_r(tmp, "/", &save_ptr);
+	while (path_element != NULL) {
+		struct node *node = lookup_node(f, ino, path_element);
+		if (node == NULL) {
+			err = -ENOENT;
+			break;
+		}
+		ino = node->nodeid;
+		path_element = strtok_r(NULL, "/", &save_ptr);
+	}
+	pthread_mutex_unlock(&f->lock);
+	free(tmp);
+
+	if (!err)
+		*inop = ino;
+	return err;
 }
 
 static char *add_name(char **buf, unsigned *bufsize, char *s, const char *name)
@@ -2509,7 +2541,8 @@ static void fuse_lib_init(void *data, struct fuse_conn_info *conn)
 	struct fuse *f = (struct fuse *) data;
 
 	fuse_create_context(f);
-	conn->want |= FUSE_CAP_EXPORT_SUPPORT;
+	if(conn->capable & FUSE_CAP_EXPORT_SUPPORT)
+		conn->want |= FUSE_CAP_EXPORT_SUPPORT;
 	fuse_fs_init(f->fs, conn, &f->conf);
 }
 
@@ -4349,7 +4382,9 @@ int fuse_loop(struct fuse *f)
 	return fuse_session_loop(f->se);
 }
 
-int fuse_loop_mt(struct fuse *f, int clone_fd)
+int fuse_loop_mt_32(struct fuse *f, struct fuse_loop_config *config);
+FUSE_SYMVER(".symver fuse_loop_mt_32,fuse_loop_mt@@FUSE_3.2");
+int fuse_loop_mt_32(struct fuse *f, struct fuse_loop_config *config)
 {
 	if (f == NULL)
 		return -1;
@@ -4358,9 +4393,19 @@ int fuse_loop_mt(struct fuse *f, int clone_fd)
 	if (res)
 		return -1;
 
-	res = fuse_session_loop_mt(fuse_get_session(f), clone_fd);
+	res = fuse_session_loop_mt(fuse_get_session(f), config);
 	fuse_stop_cleanup_thread(f);
 	return res;
+}
+
+int fuse_loop_mt_31(struct fuse *f, int clone_fd);
+FUSE_SYMVER(".symver fuse_loop_mt_31,fuse_loop_mt@FUSE_3.1");
+int fuse_loop_mt_31(struct fuse *f, int clone_fd)
+{
+	struct fuse_loop_config config;
+	config.clone_fd = clone_fd;
+	config.max_idle_threads = 10;
+	return fuse_loop_mt_32(f, &config);
 }
 
 void fuse_exit(struct fuse *f)
@@ -4397,11 +4442,19 @@ int fuse_interrupted(void)
 		return 0;
 }
 
+int fuse_invalidate_path(struct fuse *f, const char *path) {
+	fuse_ino_t ino;
+	int err = lookup_path_in_cache(f, path, &ino);
+	if (err) {
+		return err;
+	}
+
+	return fuse_lowlevel_notify_inval_inode(f->se, ino, 0, 0);
+}
+
 #define FUSE_LIB_OPT(t, p, v) { t, offsetof(struct fuse_config, p), v }
 
 static const struct fuse_opt fuse_lib_opts[] = {
-	FUSE_LIB_OPT("-h",		      show_help, 1),
-	FUSE_LIB_OPT("--help",		      show_help, 1),
 	FUSE_OPT_KEY("debug",		      FUSE_OPT_KEY_KEEP),
 	FUSE_OPT_KEY("-d",		      FUSE_OPT_KEY_KEEP),
 	FUSE_LIB_OPT("debug",		      debug, 1),
@@ -4426,7 +4479,34 @@ static const struct fuse_opt fuse_lib_opts[] = {
 	FUSE_OPT_END
 };
 
-static void fuse_lib_help(void)
+static int fuse_lib_opt_proc(void *data, const char *arg, int key,
+			     struct fuse_args *outargs)
+{
+	(void) arg; (void) outargs; (void) data; (void) key;
+
+	/* Pass through unknown options */
+	return 1;
+}
+
+
+static const struct fuse_opt fuse_help_opts[] = {
+	FUSE_LIB_OPT("modules=%s", modules, 1),
+	FUSE_OPT_KEY("modules=%s", FUSE_OPT_KEY_KEEP),
+	FUSE_OPT_END
+};
+
+static void print_module_help(const char *name,
+			      fuse_module_factory_t *fac)
+{
+	struct fuse_args a = FUSE_ARGS_INIT(0, NULL);
+	if (fuse_opt_add_arg(&a, "") == -1 ||
+	    fuse_opt_add_arg(&a, "-h") == -1)
+		return;
+	printf("\nOptions for %s module:\n", name);
+	(*fac)(&a, NULL);
+}
+
+void fuse_lib_help(struct fuse_args *args)
 {
 	/* These are not all options, but only the ones that
 	   may be of interest to an end-user */
@@ -4443,36 +4523,43 @@ static void fuse_lib_help(void)
 "    -o noforget            never forget cached inodes\n"
 "    -o remember=T          remember cached inodes for T seconds (0s)\n"
 "    -o modules=M1[:M2...]  names of modules to push onto filesystem stack\n");
-}
 
-static void fuse_lib_help_modules(void)
-{
+
+	/* Print low-level help */
+	fuse_lowlevel_help();
+
+	/* Print help for builtin modules */
+	print_module_help("subdir", &fuse_module_subdir_factory);
+#ifdef HAVE_ICONV
+	print_module_help("iconv", &fuse_module_iconv_factory);
+#endif
+
+	/* Parse command line options in case we need to
+	   activate more modules */
+	struct fuse_config conf = { .modules = NULL };
+	if (fuse_opt_parse(args, &conf, fuse_help_opts,
+			   fuse_lib_opt_proc) == -1
+	    || !conf.modules)
+		return;
+	
+	char *module;
+	char *next;
 	struct fuse_module *m;
-	printf("\nModule options:\n");
-	pthread_mutex_lock(&fuse_context_lock);
-	for (m = fuse_modules; m; m = m->next) {
-		struct fuse_fs *fs = NULL;
-		struct fuse_fs *newfs;
-		struct fuse_args args = FUSE_ARGS_INIT(0, NULL);
-		if (fuse_opt_add_arg(&args, "") != -1 &&
-		    fuse_opt_add_arg(&args, "-h") != -1) {
-			printf("\n[%s]\n", m->name);
-			newfs = m->factory(&args, &fs);
-			assert(newfs == NULL);
-		}
-		fuse_opt_free_args(&args);
+
+	// Iterate over all modules
+	for (module = conf.modules; module; module = next) {
+		char *p;
+		for (p = module; *p && *p != ':'; p++);
+		next = *p ? p + 1 : NULL;
+		*p = '\0';
+
+		m = fuse_get_module(module);
+		if (m)
+			print_module_help(module, &m->factory);
 	}
-	pthread_mutex_unlock(&fuse_context_lock);
 }
 
-static int fuse_lib_opt_proc(void *data, const char *arg, int key,
-			     struct fuse_args *outargs)
-{
-	(void) arg; (void) outargs; (void) data; (void) key;
-
-	/* Pass through unknown options */
-	return 1;
-}
+				      
 
 static int fuse_init_intr_signal(int signum, int *installed)
 {
@@ -4595,7 +4682,11 @@ void fuse_stop_cleanup_thread(struct fuse *f)
 	}
 }
 
-struct fuse *fuse_new(struct fuse_args *args,
+
+/* Explicit prototype to prevent compiler warnings
+   (fuse.h only defines fuse_new()) */
+FUSE_SYMVER(".symver fuse_new_31,fuse_new@@FUSE_3.1");
+struct fuse *fuse_new_31(struct fuse_args *args,
 		      const struct fuse_operations *op,
 		      size_t op_size, void *user_data)
 {
@@ -4620,20 +4711,15 @@ struct fuse *fuse_new(struct fuse_args *args,
 			   fuse_lib_opt_proc) == -1)
 		goto out_free;
 
-	if (f->conf.show_help) {
-		fuse_lib_help();
-		fuse_lowlevel_help();
-		/* Defer printing module help until modules
-		   have been loaded */
-	}
-
 	pthread_mutex_lock(&fuse_context_lock);
 	static int builtin_modules_registered = 0;
 	/* Have the builtin modules already been registered? */
 	if (builtin_modules_registered == 0) {
 		/* If not, register them. */
 		fuse_register_module("subdir", fuse_module_subdir_factory, NULL);
+#ifdef HAVE_ICONV
 		fuse_register_module("iconv", fuse_module_iconv_factory, NULL);
+#endif
 		builtin_modules_registered= 1;
 	}
 	pthread_mutex_unlock(&fuse_context_lock);
@@ -4671,11 +4757,6 @@ struct fuse *fuse_new(struct fuse_args *args,
 			    fuse_push_module(f, module, args) == -1)
 				goto out_free_fs;
 		}
-	}
-
-	if(f->conf.show_help) {
-		fuse_lib_help_modules();
-		goto out_free_fs;
 	}
 
 	if (!f->conf.ac_attr_timeout_set)
@@ -4753,6 +4834,32 @@ out_free:
 	free(f);
 out:
 	return NULL;
+}
+
+/* Emulates 3.0-style fuse_new(), which processes --help */
+struct fuse *fuse_new_30(struct fuse_args *args, const struct fuse_operations *op,
+			 size_t op_size, void *private_data);
+FUSE_SYMVER(".symver fuse_new_30,fuse_new@FUSE_3.0");
+struct fuse *fuse_new_30(struct fuse_args *args,
+			 const struct fuse_operations *op,
+			 size_t op_size, void *user_data)
+{
+	struct fuse_config conf;
+	const struct fuse_opt opts[] = {
+		FUSE_LIB_OPT("-h", show_help, 1),
+		FUSE_LIB_OPT("--help", show_help, 1),
+		FUSE_OPT_END
+	};
+
+	if (fuse_opt_parse(args, &conf, opts,
+			   fuse_lib_opt_proc) == -1)
+		return NULL;
+
+	if (conf.show_help) {
+		fuse_lib_help(args);
+		return NULL;
+	} else
+		return fuse_new_31(args, op, op_size, user_data);
 }
 
 void fuse_destroy(struct fuse *f)