+++ lib/helper.c
@@ -42,10 +42,13 @@ static const struct fuse_opt fuse_helper_opts[] = {
 	FUSE_HELPER_OPT("-f",		foreground),
 	FUSE_HELPER_OPT("-s",		singlethread),
 	FUSE_HELPER_OPT("fsname=",	nodefault_subtype),
-	FUSE_HELPER_OPT("subtype=",	nodefault_subtype),
 	FUSE_OPT_KEY("fsname=",		FUSE_OPT_KEY_KEEP),
+#ifndef __FreeBSD__
+	FUSE_HELPER_OPT("subtype=",	nodefault_subtype),
 	FUSE_OPT_KEY("subtype=",	FUSE_OPT_KEY_KEEP),
+#endif
 	FUSE_HELPER_OPT("clone_fd",	clone_fd),
+	FUSE_HELPER_OPT("max_idle_threads=%u", max_idle_threads),
 	FUSE_OPT_END
 };
 
@@ -130,7 +133,9 @@ void fuse_cmdline_help(void)
 	       "    -f                     foreground operation\n"
 	       "    -s                     disable multi-threaded operation\n"
 	       "    -o clone_fd            use separate fuse device fd for each thread\n"
-	       "                           (may improve performance)\n");
+	       "                           (may improve performance)\n"
+	       "    -o max_idle_threads    the maximum number of idle worker threads\n"
+	       "                           allowed (default: 10)\n");
 }
 
 static int fuse_helper_opt_proc(void *data, const char *arg, int key,
@@ -161,10 +166,13 @@ static int fuse_helper_opt_proc(void *data, const char *arg, int key,
 	}
 }
 
+/* Under FreeBSD, there is no subtype option so this
+   function actually sets the fsname */
 static int add_default_subtype(const char *progname, struct fuse_args *args)
 {
 	int res;
 	char *subtype_opt;
+
 	const char *basename = strrchr(progname, '/');
 	if (basename == NULL)
 		basename = progname;
@@ -176,7 +184,11 @@ static int add_default_subtype(const char *progname, struct fuse_args *args)
 		fprintf(stderr, "fuse: memory allocation failed\n");
 		return -1;
 	}
+#ifdef __FreeBSD__
+	sprintf(subtype_opt, "-ofsname=%s", basename);
+#else
 	sprintf(subtype_opt, "-osubtype=%s", basename);
+#endif
 	res = fuse_opt_add_arg(args, subtype_opt);
 	free(subtype_opt);
 	return res;
@@ -186,12 +198,17 @@ int fuse_parse_cmdline(struct fuse_args *args,
 		       struct fuse_cmdline_opts *opts)
 {
 	memset(opts, 0, sizeof(struct fuse_cmdline_opts));
+
+	opts->max_idle_threads = 10;
+
 	if (fuse_opt_parse(args, opts, fuse_helper_opts,
 			   fuse_helper_opt_proc) == -1)
 		return -1;
 
-	/* If neither -o subtype nor -o fsname are specified,
-	   set subtype to program's basename */
+	/* *Linux*: if neither -o subtype nor -o fsname are specified,
+	   set subtype to program's basename.
+	   *FreeBSD*: if fsname is not specified, set to program's
+	   basename. */
 	if (!opts->nodefault_subtype)
 		if (add_default_subtype(args->argv[0], args) == -1)
 			return -1;
@@ -272,18 +289,15 @@ int fuse_main_real(int argc, char *argv[], const struct fuse_operations *op,
 		goto out1;
 	}
 
-	/* Re-add --help for later processing by fuse_new()
-	   (that way we also get help for modules options) */
 	if (opts.show_help) {
-		if(args.argv[0] != '\0')
+		if(args.argv[0][0] != '\0')
 			printf("usage: %s [options] <mountpoint>\n\n",
 			       args.argv[0]);
 		printf("FUSE options:\n");
 		fuse_cmdline_help();
-		if (fuse_opt_add_arg(&args, "--help") == -1) {
-			res = 1;
-			goto out1;
-		}
+		fuse_lib_help(&args);
+		res = 0;
+		goto out1;
 	}
 
 	if (!opts.show_help &&
@@ -294,10 +308,9 @@ int fuse_main_real(int argc, char *argv[], const struct fuse_operations *op,
 	}
 
 
-	/* --help is processed here and will result in NULL */
-	fuse = fuse_new(&args, op, op_size, user_data);
+	fuse = fuse_new_31(&args, op, op_size, user_data);
 	if (fuse == NULL) {
-		res = opts.show_help ? 0 : 1;
+		res = 1;
 		goto out1;
 	}
 
@@ -319,8 +332,12 @@ int fuse_main_real(int argc, char *argv[], const struct fuse_operations *op,
 
 	if (opts.singlethread)
 		res = fuse_loop(fuse);
-	else
-		res = fuse_loop_mt(fuse, opts.clone_fd);
+	else {
+		struct fuse_loop_config loop_config;
+		loop_config.clone_fd = opts.clone_fd;
+		loop_config.max_idle_threads = opts.max_idle_threads;
+		res = fuse_loop_mt(fuse, &loop_config);
+	}
 	if (res)
 		res = 1;