+++ test/test_syscalls.c
@@ -792,6 +792,7 @@ static int test_create_unlink(void)
 	return 0;
 }
 
+#ifndef __FreeBSD__
 static int test_mknod(void)
 {
 	int err = 0;
@@ -824,6 +825,7 @@ static int test_mknod(void)
 	success();
 	return 0;
 }
+#endif
 
 #define test_open(exist, flags, mode)  do_test_open(exist, flags, #flags, mode)
 
@@ -1280,6 +1282,7 @@ static int test_rename_dir(void)
 	return 0;
 }
 
+#ifndef __FreeBSD__
 static int test_mkfifo(void)
 {
 	int res;
@@ -1311,6 +1314,7 @@ static int test_mkfifo(void)
 	success();
 	return 0;
 }
+#endif
 
 static int test_mkdir(void)
 {
@@ -1329,7 +1333,9 @@ static int test_mkdir(void)
 	if (res == -1)
 		return -1;
 	err += check_mode(testdir, 0755);
-	err += check_nlink(testdir, 2);
+	/* Some file systems (like btrfs) don't track link
+	   count for directories */
+	//err += check_nlink(testdir, 2);
 	err += check_dir_contents(testdir, dir_contents);
 	res = rmdir(testdir);
 	if (res == -1) {
@@ -1445,11 +1451,13 @@ int main(int argc, char *argv[])
 
 	err += test_create();
 	err += test_create_unlink();
-	err += test_mknod();
 	err += test_symlink();
 	err += test_link();
 	err += test_link2();
+#ifndef __FreeBSD__	
+	err += test_mknod();
 	err += test_mkfifo();
+#endif
 	err += test_mkdir();
 	err += test_rename_file();
 	err += test_rename_dir();